requirements.txt
beautifulsoup4==4.9.3
config==0.5.0.post0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.8
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-model-utils==4.0.0
django-registration==3.1.1
django-star-ratings==0.9.1
django-tinymce==3.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
importlib-metadata==2.0.0
Pillow==8.0.1
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
swapper==1.1.2.post1
whitenoise==5.2.0
zipp==3.3.2
